+++ deepdelve/__init__.py
@@ -2,7 +2,7 @@
 
 from .deepdelve import DeepDelve
 
-__version__ = "4.3.1"
+__version__ = "4.3.2"
 
 
 async def setup(bot):